Added files via upload
All basic functionality in the preferences tab should work, but the commands to be sent serially need to be uncommented from the code.  The "Purge" buttons require the two new switch cases to be added to the Arduino firmware, and are based on the S0, S255 scheme we messaged about last night.  Seems to work right.  So, we will be able to show Dr. Dai purging, changing pressures digitally, changing valve opening times digitally, and I hope we will be able to show him squares and circles, and freeform if you've gotten that far.